feat(actions): cd to dirs when using west
When specifying a custom `config_path`, the follow-up message from `west
init` is to run `west update` from inside its parent directory, which
the reusable workflow didn't do before this commit. The other steps in
the workflow also seems to assume that they are within this directory,
so this makes sure that we are.

Additionally, use `GITHUB_WORKSPACE` for renaming the build artifacts,
since this path is absolute.